BE/requirements.txt
# Minimal Backend Requirements
jaclang>=0.7.0
byllm>=0.1.0
google-generativeai>=0.3.0
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
gitpython>=3.1.0
python-dotenv>=1.0.0
jac-cloud